refactor(nat): consolidate service RPC and share value models
Collapse the four-module `xr_ai_nat.functions._rpc` package into a single
`xr_ai_nat.functions._service.rpc` module (private `_pack`/`_unpack`/
`_PROTOCOL_VERSION`; unchanged wire format), and move the shared coordinate
value models (`Vector3`, `SpatialFrame`, plus `Color` and the `ServiceResult`
base) out of `spatial_math/schemas.py` into a capability-neutral
`xr_ai_nat.functions.types`. All importers (nat capabilities, the openxr and
video-memory services, the sample scene service, the video MCP adapter, and
tests) point at the new locations.

Behaviour-preserving: the RPC protocol and the models' JSON string rendering
are identical; this is an ownership/location refactor toward one home per
shared concern. No dependency changes.
