+++ agent-mcp-servers/video-mcp/video_mcp_server/__main__.py
@@ -15,7 +15,7 @@ from loguru import logger
 from xr_ai_agent import FrameUnavailable, LiveFrameSource, ProcessorEndpoint, Subscribe
 
 from xr_ai_logging import setup_logging
-from xr_ai_nat.functions._rpc import RPCError
+from xr_ai_nat.functions._service.rpc import RPCError
 from xr_ai_nat.functions.video_memory._client import VideoMemoryClient
 from xr_ai_nat.functions.video_memory.schemas import (
     HistoricalFrameRequest,
