+++ agent-samples/xr-render-demo/scene/xr_render_scene/client.py
@@ -3,7 +3,7 @@
 
 """Typed client for the sample-local scene process."""
 
-from xr_ai_nat.functions._rpc import RPCClient
+from xr_ai_nat.functions._service.rpc import RPCClient
 
 from .schemas import (
     AddPrimitiveRequest,
