--- agent-sdk/xr-ai-nat/xr_ai_nat/functions/_rpc/__init__.py
@@ -1,10 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Private transport used by service-backed XR functions."""
-
-from .client import RPCClient
-from .protocol import RPCError
-from .server import RPCServer
-
-__all__ = ["RPCClient", "RPCError", "RPCServer"]
