+++ agent-sdk/xr-ai-nat/xr_ai_nat/functions/_service/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Private transport and value models shared by service-backed functions."""
