+++ agent-sdk/xr-ai-nat/xr_ai_nat/functions/spatial_math/_math.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import math
 from typing import Literal
 
-from .schemas import SpatialFrame, Vector3
+from ..types import SpatialFrame, Vector3
 
 
 def _position(x: float, y: float, z: float) -> dict[str, float]:
