--- agent-sdk/xr-ai-nat/xr_ai_nat/functions/spatial_math/schemas.py
@@ -1,34 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Transport-neutral values used by spatial-math functions."""
-
-from __future__ import annotations
-
-from pydantic import BaseModel
-
-
-class _StructuredValue(BaseModel):
-    def __str__(self) -> str:
-        # Text-only agent bridges call str(); JSON keeps the value reusable as a typed tool input.
-        return self.model_dump_json()
-
-
-class Vector3(_StructuredValue):
-    """A position or direction in three-dimensional space."""
-
-    x: float
-    y: float
-    z: float
-
-
-class SpatialFrame(_StructuredValue):
-    """An origin and orthogonal directions supplied by a tracking capability."""
-
-    origin: Vector3
-    forward: Vector3
-    right: Vector3
-    up: Vector3
-
-
-__all__ = ["SpatialFrame", "Vector3"]
