+++ services/video-memory-service/video_memory_service/store.py
@@ -7,7 +7,7 @@ import json
 from pathlib import Path
 
 from loguru import logger
-from xr_ai_nat.functions._rpc import RPCError
+from xr_ai_nat.functions._service.rpc import RPCError
 
 
 def safe_name(value: str) -> str:
